2020: Julia day 1b
Changed getPairSum to 'return nothing' instead of throwing an error, as
it feels cleaner to test if return is nothing rather than catching an
exception (since the exception is the common path, rather than
exceptional!). However, I don't know if this is the idiomatic way - both
whether I should just use exceptions, and because it feels rather unsafe
compared to, say Haskell's 'case' on a Maybe. Perhaps this is just part
of the deal of using a dynamic language.

We also explicitly 'print' our answers, as the previous code relied on
the last expression in a script being implicitly printed - 'day1a(...)'
is not the last line anymore!

I am finding it difficult to find good documentation for Julia's
libraries -- what is a 'Set' (BST?, HashTable?), and what is the
complexity of building one from a vector?