2020: Julia day 1: use === nothing
According to the documentation
https://docs.julialang.org/en/v1/manual/performance-tips/#Checking-for-equality-with-a-singleton
it can be better for performance to use "identicality" === rather than
equality == when comparing with a singleton. This performance benefits
are completely trivial here, but let's get into the habit of good style.

Talking of style, it appears that testing '!== nothing' and then pattern
matching a pair is decent style (I wondered about this in the previous
commit message).
From the documentation on iteration
https://docs.julialang.org/en/v1/base/collections/#lib-collections-iteration
we have that a 'for' loop is translated into
  next = iterate(iter)
  while next !== nothing
      (i, state) = next
      # body
      next = iterate(iter, state)
  end