2020: Julia day 6b
Same as part a, but intersection rather than union. This made it
slightly more awkward syntactically, as I wanted to take the
intersection using Sets, rather than Vectors, and the identity element
is thus the full set rather than the empty set. To avoid this, let's
just start off with the Set constructed from the first person, which
leads to the syntactical awkwardness.