2020: Julia day 7b via linear algebra
Using floating-point arithmetic, the speed is comparable to our search,
but computes all-pairs-any-depth-containment. Most of the time in day7b2
goes into building the matrix M in the first place, rather than
inverting matrices. However, we do compute an approximation to the
answer, as we use floats to compute an integer.

If we change to using rationals, matrix inversion is suddenly much much
slower (20 seconds rather than less than 0.1 seconds), but we do get the
exact correct answer!

This was the first time I really tripped up with mutation - the new
method was returning garbage, and it turned out that day7b was mutating
its input. Easy to fix with deepcopy when you realise, but I do miss
pure languages!