2020: Julia day 14b
I am surprised that the brute force approach works so well today!

We slightly generalise the reading of the input data to handle the case
where a mask has no zeros or no ones. This was not the case for my data,
but was for the example data I used for debugging.