2021: racket day 19b
Thankfully we can re-use the work from part a to solve part b. We take
advantage of this and solve both parts together today.

There are some fairly straightforward algorithmic improvements available
(for a scanner, consider the set {|v-w| : v,w <- beacons} and prioritise
stitching the scan based on overlap of these distance-sets, note that
this doesn't require considering the 24 orientations), but we do not
implement these.