2021: racket day 22b
I wasted a whole bunch of time trying to keep track of positions using a
quadtree, but it was too slow and memory-hungry (e.g. cuboids with high
"eccentricity" (i.e. long in one or two dimensions and short in the
other(s)) require a lot of subdivision). This solution is so much
cleaner!